pyproject.toml
[project]
name = "cloudwalk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-google-vertexai>=3.2.0",
    "langgraph>=0.2.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typing-extensions>=4.0.0",
]
